Add Id field to UpdateItemCommand class
Added an Id property to the UpdateItemCommand class. Replaced the retrieval and validation of items with the Id property instead of ItemCode. The change was made because using an Id is a more reliable way to specifically identify items in the database. The item retrieval and validation have been updated accordingly, plus the Update method now accepts the Id as parameter for more accurate data handling.