"Add IsActive status and new models to Client Prospecting Feature"
In the ProspectingMappingProfiles, GetAllRequestedProspectAsync, GetRequestedProspectById, GetAllRejectProspectAsync, GetAllApprovedProspectAsync, and GetRequestedProspectById sections, IsActive entry was added for better client activity monitoring. The status of the clients whether they are active or not, is now part of our detailed mapping.

In the DataContext, new entities of approved and rejected freebies were also introduced. These modifications are needed to handle the expanded business logic related to the management of freebies which comes with an approval or rejection process.

The IsActive field and mentioned models implemented provide more efficient data extraction and improve the efficiency of the response, considering the application's evolving needs and demand.