"Remove InitialCreate migration file"
This commit removes the InitialCreate.cs from the RDF.Arcana.API/Migrations folder. This migration file generated code for creating all the tables for the initial app model. It was deleted since the model and its requirements have changed dramatically, so a new initial migration will be created to accurately reflect the new state and relationships of the models in the app. This also ensures the app's database is in sync with the current model.