"Add 'Reason' and 'StoreType' fields to client prospecting feature
The 'Reason' field was added to the 'UpdateRejectProspectStatus', 'UpdateApprovedProspectStatus', 'UpdateRequestProspectStatus', and 'RejectProspectRequest' classes. This is to keep a record of the reason why a prospect status was updated.

A 'StoreType' field is added to the 'AddNewProspectCommand' class to allow identifying the type of store that is being added as a new prospect. For usability and data tracking purposes, this additional information has become necessary."