Add 'Reason' field to prospecting entities
A 'Reason' field was added to the RecordClient, GetAllRequestedProspectResult, GetAllRejectProspectResult entities. This field stores the reason why a prospect status was updated, thus providing more context for status changes. Furthermore, a filter was added in 'Handle' method of GetAllRequestedProspectQuery and GetAllRejectProspectQuery, to only get records where 'IsActive' is true. This change makes sure that only relevant data is fetched.