Refined status checks in prospect feature
Enhanced the status checks in the prospecting feature for more accurate queries. Changed the attribute in the 'UpdateRejectProspectStatus.cs' file from 'ApprovalType' to 'Client.RegistrationStatus' in order to specifically check if the client's registration is rejected. The 'GetAllRequestedProspectAsync.cs' file has been updated to include an additional condition to the `requestedProspect` query. The query now also checks if the `Client.RegistrationStatus` is 'Requested' along with other conditions. All these changes are designed to ensure only relevant results are considered during the queries process, enhancing the overall efficiency of the prospecting feature.