"Add functionality to update approved prospect
A new functionality was added to update an approved prospect using `UpdateApprovedProspectCommand`. This feature was added to `UpdateApprovedProspect.cs` file in the prospecting features. This new functionality enables updating the details of any approved prospect.

Also, file renaming took place for `UpdateProspectRequest.cs` to `UpdateProspectInformation.cs` for aligning the function name with its purpose, providing up-to-date and accurate information of the prospect. And a new condition was added to check for 'Approver Approval' when a client is being updated.

These changes allow managing approved prospects more effectively, ensuring data consistency, readability, and maintainability."