"Add new features, enhance existing ones and refactor related files
Added UpdateFreebiesInformation feature to allow updating freebies. This not only enhances the functionality of the application but, also grants the flexibility to manipulate freebie data efficiently for users. Created UserRoleAlreadyTaggedException for error handling. Also, included 'User' field in some queries in GetUserRolesAsync. Enhanced IsTagged check and included the refaction in UserRoleMappingExtension. Removed unnecessary 'Domain' namespace from GetAllRequestedProspectAsync. In AddNewUSer.cs, added a new additional check to throw UserRoleAlreadyTaggedException if the user role is already in use. These amendments were a necessity to deliver extended functionality, more readable and reliable data to users, with the commitment to data consistency."