Update Freebie management and Prospecting enhancements
Updated the Freebies management by allowing the removal of items from a Freebie request. Added a new status "Released" for Prospecting and improved the query result details for approved Freebies.

These changes allow the application to track the status of Prospecting and the Freebie requests in a more detailed and efficient manner. Now it is possible to remove items from a Freebie request. For Prospecting, a new status "Released" was introduced to clearly distinguish the delivery status of the client. On top of that, the return details for approved Freebies was enhanced to include the FreebieRequestId and ClientId, providing more visibility and control for the management.