"Removed Modules entity and refactored UserRole permissions
In this commit, the 'Modules' entity has been removed as it was found to be unnecessary. The permissions system within the 'UserRoles' entity has been refactored from being a collection of 'ModuleId' to a collection of strings. This provides the same functionality with a simpler implementation. Other classes which used the 'Modules' class such as 'GetUsersAsync' and 'AuthenticateUser' have been updated to reflect these changes as well. The file 'Module.cs' was deleted."