Removed Modules entity and refactored UserRole permissions
Modules entity was determined to be unnecessary and was therefore removed. The permissions within the UserRoles entity have been refactored, changing from a collection of ModuleId to a collection of strings. This streamlines the implementation while maintaining the same functionality. All classes that utilized the Modules entity, such as GetUsersAsync and AuthenticateUser, have been updated accordingly.