Update Freebies information and refactor permissions
Modified the UpdateFreebiesInformation class to add RequestId to the FreebieItems. This detail offers more information about the freebies request, helping in data tracking and management. Also, refactored permission system for UserRoles. The previously used 'Modules' entity was found superfluous and replaced with a collection of strings. This change simplifies implementation and maintains functionality.