Add latitude and longitude fields to clients table
Added latitude and longitude columns to the clients table in order to store client's geographical locations. This will provide more accurate data for mapping and geolocation services. The migration contains the Up and Down methods for adding and removing the fields respectively. The file 'AddPinLocation.Designer.cs' is also updated to reflect these changes.

Signed-off-by: aldrinvega <aldrinmundovega@gmail.com>