Update DirectRegistration and clean workspace.xml files
The workspace.xml related to .idea.RDF.Arcana workspace was cleaned removing unused and redundant tracks of changes committed and component values. All the obsolete changes tracked which are no longer relevant and redundant keys which were interfering with the workspace's readability were removed, making it less cluttered and more maintainable.

Additionally, the AddTermAndCondition script was updated in the Prospecting module, changing BookingCoverage to BookingCoverageId in term options for precise referencing. This change corrects the variable name and adds more clarity to its purpose. Meanwhile, in the DirectRegistration script the value of 'PROSPECT_TYPE' was changed to 'DIRECT' enhancing the data's consistency and reflecting its actual use. The feature 'GetAllDirectRegistrationClients' was also added to return paginated client data from direct

Signed-off-by: aldrinvega <aldrinmundovega@gmail.com>