Adjust foreign key relationship for approvals
Modified the foreign key relationship for approvals in the freebie_requests table to ensure referential integrity. This was done by first dropping and renaming the 'approval_id' foreign key to 'approvals_id'. Then, the corresponding index was created for 'approvals_id', and a new foreign key was added. These change was necessary to maintain the correct relationship between the approval and freebie_requests tables during database operations.

Signed-off-by: aldrinvega <aldrinmundovega@gmail.com>